refactor: refactor remoteDataSourceModule and rename RemoteDataSource to RemoteSearchDataSource
